Fix warnings when building with Visual Studio.

+++ include/ortp/rtcp.h
@@ -526,12 +526,12 @@ typedef struct OrtpLossRateEstimator{
 	* Total number of outgoing duplicate packets on last
 	* ortp_loss_rate_estimator_process_report_block iteration.
 	**/
-	int32_t last_dup_packet_sent_count;
+	int64_t last_dup_packet_sent_count;
 	/**
 	* Total number of outgoing unique packets on last
 	* ortp_loss_rate_estimator_process_report_block iteration.
 	**/
-	int32_t last_packet_sent_count;
+	int64_t last_packet_sent_count;
 }OrtpLossRateEstimator;
